backend:dislike fonction frontend:compenent comment

+++ backend/controlles/comment.js
@@ -24,18 +24,17 @@ exports.createComment = (req, res, next) => {
     .catch((error) => res.status(500).json({ error }));
 };
 exports.updateComment = (req, res, next) => {
-  models.post
-    .findOne({ where: { id: req.params.postId } })
-    .then((post) => {
-      console.log(post);
-      models.comment
-        .findOne({ where: { userId: req.auth.userId } })
-        .then((comment) => {
-          console.log(comment);
-          if (comment.userId !== req.auth.userId) {
-            return res.status(401).json({ error: "unauthorized request" });
-          }
-          if (req.file) {
+  models.post.findOne({ where: { id: req.params.postId } }).then((post) => {
+    console.log(post);
+    models.comment
+      .findOne({ where: { userId: req.auth.userId } })
+      .then((comment) => {
+        console.log(comment);
+        if (comment.userId !== req.auth.userId) {
+          return res.status(401).json({ error: "unauthorized request" });
+        }
+        if (req.file) {
+          if (comment.image_url != null) {
             const filename = comment.image_url.split("/images")[1]; //le nom de l'image a supprimer
             fs.unlink(`images/${filename}`, function (error) {
               //supression de l'ancienne image
@@ -44,33 +43,49 @@ exports.updateComment = (req, res, next) => {
               }
             });
           }
-          const commentObject = req.file
-            ? {
-                content: req.body.content,
-                image_url: `${req.protocol}://${req.get("host")}/images/${
-                  req.file.filename
-                }`,
-              }
-            : { content: req.body.content };
-          comment
-            .update({ ...commentObject }, { where: { id: req.params.id } })
-            .then(() => {
-              res.status(201).json({ message: "Comment edit !" });
-            })
-            .catch((error) => res.status(400).json({ error }));
-        })
-        .catch((error) => res.status(500).json({ error }));
-    })
-    .catch((error) => res.status(500).json({ error }));
+        }
+        const commentObject = req.file
+          ? {
+              content: req.body.content,
+              image_url: `${req.protocol}://${req.get("host")}/images/${
+                req.file.filename
+              }`,
+            }
+          : { content: req.body.content };
+        comment
+          .update({ ...commentObject }, { where: { id: req.params.id } })
+          .then(() => {
+            res.status(201).json({ message: "Comment edit !" });
+          })
+          .catch((error) => res.status(400).json({ error }));
+      });
+    //.catch((error) => res.status(500).json({ error }));
+  });
+  //.catch((error) => res.status(500).json({ error }));
 };
 exports.deleteComment = (req, res, next) => {
   models.comment
     .findOne({ where: { postId: req.params.postId, id: req.params.id } })
     .then((comment) => {
+      console.log(comment.userId);
+      console.log(req.auth.userId);
       if (comment.userId == req.auth.userId || req.auth.isAdmin == 1) {
         //nom du fichier à supprimer
-        const filename = comment.image_url.split("/images")[1];
-        fs.unlink(`images/${filename}`, () => {
+        if (comment.image_url != null) {
+          const filename = comment.image_url.split("/images")[1];
+          fs.unlink(`images/${filename}`, () => {
+            comment
+              .destroy({
+                where: { postId: req.params.postId, id: req.params.id },
+              })
+              .then(() => {
+                res.status(201).json({ message: "Comment delete !" });
+              })
+              .catch((error) => {
+                res.status(400).json({ error });
+              });
+          });
+        } else {
           comment
             .destroy({
               where: { postId: req.params.postId, id: req.params.id },
@@ -81,12 +96,12 @@ exports.deleteComment = (req, res, next) => {
             .catch((error) => {
               res.status(400).json({ error });
             });
-        });
+        }
       } else {
         return res.status(401).json({ message: "Unauthorized request" });
       }
-    })
-    .catch((error) => res.status(500).json({ error }));
+    });
+  //.catch((error) => res.status(500).json({ error:"le commentaire n'existe pas dans la bd" }));
 };
 exports.readComment = (req, res, next) => {
   models.comment
